+++ site/src/tutorial.rs
@@ -959,13 +959,6 @@ splitArray([1, 2, 3, 7, 2, 4, 5])"</code>
         <p>"Because a second "<code>"|"</code>" immediately after another indicates a signature, branches that do nothing must contain "<Prim prim=Identity/>"."</p>
         <Editor example="f ← (+5|∘|÷10)+∩>5,10.\n[f2 f6 f200]"/>
 
-        <h2 id="if"><Prim prim=If/></h2>
-        <p>"The "<Prim prim=If/>" modifier is similar to a switch function, but it only has two branches. The true branch is the first function and the false branch is the second"</p>
-        <Editor example="f ← ?+×\nf 0 3 5\nf 1 3 5"/>
-        <p>"This may seem useless when switch functions exist, and for scalar conditions, apart from being slightly shorter, it is. However, "<Prim prim=If/>"'s condition can be a list."</p>
-        <Editor example="?+- [1 0 1] [2 2 2] [4 4 4]"/>
-        <p>"While "<Prim prim=If/>" can be chained, it is usually preferable to use a switch function instead."</p>
-
         <h2 id="assert"><Prim prim=Assert/></h2>
         <p>"The "<Prim prim=Assert/>" function takes any value and a condition. If the condition is anything but "<code>"1"</code>", the value is thrown as an error that can be caught with "<Prim prim=Try/>"."</p>
         <Editor example="f ← ⍣(¯⍤10≤10.);\nf 5\nf 12"/>