+++ src/algorithm/fork.rs
@@ -1,6 +1,10 @@
 //! Algorithms for forking modifiers
 
-use crate::{value::Value, Uiua, UiuaResult};
+use std::sync::Arc;
+
+use crate::{value::Value, Function, Shape, Signature, Uiua, UiuaResult};
+
+use super::multi_output;
 
 pub fn both(env: &mut Uiua) -> UiuaResult {
     let f = env.pop_function()?;
@@ -75,133 +79,82 @@ pub fn bracket(env: &mut Uiua) -> UiuaResult {
     Ok(())
 }
 
-pub fn iff(env: &mut Uiua) -> UiuaResult {
-    let if_true = env.pop_function()?;
-    let if_false = env.pop_function()?;
-    let condition = env.pop(1)?;
-    if let Ok(condition) = condition.as_nat(env, "") {
-        if condition > 1 {
-            return Err(env.error(format!(
-                "If's condition must be 0 or 1, but it is {}",
-                condition
-            )));
-        }
-        let if_true_sig = if_true.signature();
-        let if_false_sig = if_false.signature();
-        if if_true_sig.args == if_false_sig.args || if_true_sig.is_compatible_with(if_false_sig) {
-            if condition == 1 {
-                env.call(if_true)?;
-            } else {
-                env.call(if_false)?;
-            }
-        } else {
-            let arg_count = if_true_sig.args.max(if_false_sig.args);
-            let mut args = Vec::with_capacity(arg_count);
-            for i in 0..arg_count {
-                args.push(env.pop(i + 1)?);
-            }
-            if condition == 1 {
-                for arg in args.into_iter().take(if_true_sig.args).rev() {
-                    env.push(arg);
-                }
-                env.call(if_true)?;
-            } else {
-                for arg in args.into_iter().take(if_false_sig.args).rev() {
-                    env.push(arg);
-                }
-                env.call(if_false)?;
-            }
-        }
+pub fn switch(count: usize, sig: Signature, env: &mut Uiua) -> UiuaResult {
+    // Get selector
+    let selector = env
+        .pop("switch index")?
+        .as_natural_array(env, "Switch index must naturals")?;
+    if let Some(i) = selector.data.iter().find(|&&i| i >= count) {
+        return Err(env.error(format!(
+            "Switch index {i} is out of bounds for switch of size {count}"
+        )));
+    }
+    // Switch
+    if selector.rank() == 0 {
+        // Scalar
+        let i = selector.data[0];
+        // Get function
+        let Some(f) = env
+            .function_stack
+            .drain(env.function_stack.len() - count..)
+            .nth(i)
+        else {
+            return Err(env.error(
+                "Function stack was empty when getting switch function. \
+            This is a bug in the interpreter.",
+            ));
+        };
+        // Discard unused arguments
+        let discard_start = env.stack.len().saturating_sub(sig.args);
+        let discard_end =
+            discard_start + sig.args - f.signature().args - (sig.outputs - f.signature().outputs);
+        env.stack.drain(discard_start..discard_end);
+        env.call(f)
     } else {
-        let condition = condition.as_nats(
-            env,
-            "If's condition must be a natural number or list of natural numbers",
-        )?;
-        if !condition.iter().all(|&x| x <= 1) {
-            return Err(env.error(format!(
-                "If's condition must be all 0s or 1s, but it is {:?}",
-                condition
-            )));
-        }
-        let if_true_sig = if_true.signature();
-        let if_false_sig = if_false.signature();
-        if if_true_sig.outputs != 1 {
-            return Err(env.error(format!(
-                "If's true branch must return 1 value, but it returns {}",
-                if_true_sig.outputs
-            )));
-        }
-        if if_false_sig.outputs != 1 {
-            return Err(env.error(format!(
-                "If's false branch must return 1 value, but it returns {}",
-                if_false_sig.outputs
-            )));
-        }
-        let arg_count = if_true_sig.args.max(if_false_sig.args);
-        match arg_count {
-            1 => {
-                let xs = env.pop(2)?;
-                if xs.row_count() != condition.len() {
-                    return Err(env.error(format!(
-                        "If's condition must have the same number of rows as its argument, \
-                        but it has {} rows and its argument has {} rows",
-                        condition.len(),
-                        xs.row_count()
-                    )));
-                }
-                let mut new_rows = Vec::with_capacity(condition.len());
-                for (con, x) in condition.into_iter().zip(xs.into_rows()) {
-                    env.push(x);
-                    if con == 1 {
-                        env.call(if_true.clone())?;
-                        new_rows.push(env.pop("if's true branch result")?);
-                    } else {
-                        env.call(if_false.clone())?;
-                        new_rows.push(env.pop("if's false branch result")?);
-                    }
-                }
-                env.push(Value::from_row_values(new_rows, env)?);
-            }
-            2 => {
-                let a = env.pop(2)?;
-                let b = env.pop(3)?;
-                if a.row_count() != condition.len() || b.row_count() != condition.len() {
-                    return Err(env.error(format!(
-                        "If's condition must have the same number of rows as its arguments, \
-                        but it has {} rows and its arguments have {} and {} rows",
-                        condition.len(),
-                        a.row_count(),
-                        b.row_count()
-                    )));
-                }
-                let mut new_rows = Vec::with_capacity(condition.len());
-                for (con, (a, b)) in condition.into_iter().zip(a.into_rows().zip(b.into_rows())) {
-                    if con == 1 {
-                        if if_true_sig.args == 2 {
-                            env.push(b);
-                        }
-                        env.push(a);
-                        env.call(if_true.clone())?;
-                        new_rows.push(env.pop("if's true branch result")?);
-                    } else {
-                        if if_false_sig.args == 2 {
-                            env.push(b);
-                        }
-                        env.push(a);
-                        env.call(if_false.clone())?;
-                        new_rows.push(env.pop("if's false branch result")?);
-                    }
-                }
-                env.push(Value::from_row_values(new_rows, env)?);
-            }
-            n => {
+        // Array
+        // Collect arguments
+        let mut args_rows: Vec<_> = Vec::with_capacity(sig.args);
+        for i in 0..sig.args {
+            let arg = env.pop(i + 1)?;
+            if !arg.shape().starts_with(selector.shape()) {
                 return Err(env.error(format!(
-                    "The maximum of iterating if's function's arguments must be 1 or 2, \
-                    but it is {n}"
+                    "The function's select's shape {} is not compatible \
+                    with the argument {}'s shape {}",
+                    selector.format_shape(),
+                    i + 1,
+                    arg.format_shape(),
                 )));
             }
+            let row_shape = Shape::from(&arg.shape()[selector.rank()..]);
+            args_rows.push(arg.into_row_shaped_slices(row_shape));
         }
+        args_rows.reverse();
+        // Collect functions
+        let functions: Vec<Arc<Function>> = env
+            .function_stack
+            .drain(env.function_stack.len() - count..)
+            .collect();
+        let mut outputs = multi_output(sig.outputs, Vec::new());
+        for elem in selector.data {
+            let f = &functions[elem];
+            for (i, arg) in args_rows.iter_mut().rev().enumerate().rev() {
+                let arg = arg.next().unwrap();
+                if i < f.signature().args {
+                    env.push(arg);
+                }
+            }
+            env.call(f.clone())?;
+            for i in 0..sig.outputs {
+                outputs[i].push(env.pop("switch output")?);
+            }
+        }
+        for output in outputs.into_iter().rev() {
+            let mut new_value = Value::from_row_values(output, env)?;
+            let mut new_shape = selector.shape.clone();
+            new_shape.extend_from_slice(&new_value.shape()[1..]);
+            *new_value.shape_mut() = new_shape;
+            env.push(new_value);
+        }
+        Ok(())
     }
-
-    Ok(())
 }