+++ changelog.md
@@ -6,7 +6,8 @@ Uiua is not yet stable.
 This version is not yet released.
 If you are reading this on the website, then these changes are live here.
 ### Language
-- Switch functions are now less strict about branch signature compatibility
+- Switch functions are now less strict about branch signature compatibility and can take arrays as conditions
+- Remove `if ?`, as all its use cases are now covered by switch functions. It will continue to parse, but `?ab` will be formatted as `(b|a)`
 - [`under` `⍜`](https://uiua.org/docs/under) now works with [`absolut value` `⌵`](https://uiua.org/docs/absolute)
 ### Interpreter
 - Lots of bug and crash fixes